ServeAPI/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pillow==11.2.1
python-multipart==0.0.20
lightning==2.5.1
lightning-utilities==0.14.3
mlflow==2.9.2
pytorch-lightning==2.5.1
rich-argparse==1.6.0
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
prometheus-client 
nvidia-ml-py3

